These enhancements were made using the following test case, which has 100 `DropdownField` instances with 100 options each:

```php
<?php

use SilverStripe\Control\Controller;
use SilverStripe\Control\HTTPRequest;
use SilverStripe\Control\Session;
use SilverStripe\Forms\DropdownField;
use SilverStripe\Forms\FieldList;
use SilverStripe\Forms\Form;

$data = [];
foreach (range(0, 100) as $i) {
    $data[$i] = "Demo $i";
}
$fields = new FieldList();
foreach (range(1, 100) as $i) {
    $fields->push(new DropdownField("Drop{$i}", null, $data));
}

// Necessary for form to render
Controller::config()->set('url_segment', '123');
$controller = new Controller();
$controller->setRequest((new HTTPRequest('GET', ''))->setSession(new Session([])));

$form = new Form(
    $controller,
    fields: $fields
);
$form->disableSecurityToken();

$start = hrtime(true);
$form->forTemplate();
$end = hrtime(true);
$timeInMs = ($end - $start) / 1e+6;
var_dump("{$timeInMs}ms");
```

|Scenario|Time to render|
|---|---|
|Without these changes|~1000ms|
|`CastingService` improvement only|~900ms|
|`ViewLayerData` improvement only|~800ms|
|Both improvements|~700ms|